Base of script

+++ halton.cpp
@@ -1,32 +1,32 @@
 #include "halton.h"
 
-double halton(int index, int base) {
+float halton(int index, int base) {
     int d;
     int j;
-    double prime_inv = 1.0 / static_cast<double>(base);
-    double result = 0.0;
+    float prime_inv = 1.0 / static_cast<float>(base);
+    float result = 0.0;
     int t = index;
 
     while (t > 0) {
         d = (t % base);
-        result += static_cast<double>(d) * prime_inv;
-        prime_inv = prime_inv / static_cast<double>(base);
+        result += static_cast<float>(d) * prime_inv;
+        prime_inv = prime_inv / static_cast<float>(base);
         t = (t - d) / base;
     }
     return result;
 }
 
-double halton_scramble(int index, int base) {
+float halton_scramble(int index, int base) {
     int d;
     int j;
-    double prime_inv = 1.0 / static_cast<double>(base);
-    double result = 0.0;
+    float prime_inv = 1.0 / static_cast<float>(base);
+    float result = 0.0;
     int t = index;
 
     while (t > 0) {
         d = (t % base);
-        result += static_cast<double>(d) * prime_inv;
-        prime_inv = prime_inv / static_cast<double>(base);
+        result += static_cast<float>(d) * prime_inv;
+        prime_inv = prime_inv / static_cast<float>(base);
         t = (t - d) / base;
     }
