+++ sobol.h
@@ -5,6 +5,6 @@
 #include "sbprimes.h"
 #include "scramble.h"
 
-double sobol(uint32_t index, uint32_t base);
+float sobol(uint32_t index, uint32_t base);
 
-double sobol_scramble(uint32_t index, uint32_t base);
+float sobol_scramble(uint32_t index, uint32_t base);
